backend/requirements.txt
# FastAPI Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9

# Database Clients
# Supabase (DEPRECATED - keeping for backwards compatibility during migration)
supabase>=2.0.0

# GCP Cloud SQL
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
cloud-sql-python-connector[pg8000]>=1.4.0

# GCP Vertex AI Vector Search (replaces Pinecone)
google-cloud-aiplatform>=1.38.0

# Pinecone (DEPRECATED - keeping for backwards compatibility)
pinecone-client>=3.0.0

# Neo4j
neo4j>=5.0.0

# AI/LLM
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
pdfplumber>=0.11.0
pdf2image>=1.16.0
PyMuPDF>=1.23.0

# Background Jobs
inngest>=0.3.0

# MLflow for AI/LLM Tracking & Monitoring
mlflow>=2.9.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
